Move Custom Control to new project
