+++ src/WPFApplication/App.xaml
@@ -5,7 +5,17 @@
              StartupUri="MainWindow.xaml">
   <Application.Resources>
     <ResourceDictionary>
- 
+      <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary Source="pack://application:,,,/WPFCustomControlLib;component/Themes/Button.xaml"/>
+        <ResourceDictionary Source="pack://application:,,,/WPFCustomControlLib;component/Themes/ListView.xaml"/>
+      </ResourceDictionary.MergedDictionaries>
+
+      <Style x:Key="appWindow" TargetType="{x:Type Window}">
+        <Setter Property="FontSize" Value="14"/>
+        <Setter Property="Height" Value="450"/>
+        <Setter Property="Width" Value="800"/>
+      </Style>
+
       <Style x:Key="appViewTitleStyle" TargetType="{x:Type TextBlock}">
         <Setter Property="FontSize" Value="24"/>
         <Setter Property="Margin" Value="0"/>
