+++ src/WPFApplication/Properties/AssemblyInfo.cs
@@ -1,7 +1,7 @@
 ﻿using System.Runtime.InteropServices;
 using System.Windows;
 
-[assembly: ThemeInfo( System.Windows.ResourceDictionaryLocation.None, ResourceDictionaryLocation.SourceAssembly )]
+[assembly: ThemeInfo( ResourceDictionaryLocation.None, ResourceDictionaryLocation.SourceAssembly )]
 
 // In SDK-style projects such as this one, several assembly attributes that were historically
 // defined in this file are now automatically added during build and populated with
