--- src/WPFApplication/Themes/Generic.xaml
@@ -1,60 +1,0 @@
-﻿<ResourceDictionary
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="clr-namespace:WPFApplication">
-
-  <SolidColorBrush x:Key="TextBox.Static.Border" Color="#FFABAdB3"/>
-  <SolidColorBrush x:Key="TextBox.MouseOver.Border" Color="#FF7EB4EA"/>
-  <SolidColorBrush x:Key="TextBox.Focus.Border" Color="#FF569DE5"/>
-
-  <Style TargetType="{x:Type local:CustomControl1}">
-    <Setter Property="Background" Value="{DynamicResource {x:Static SystemColors.WindowBrushKey}}"/>
-    <Setter Property="BorderBrush" Value="{StaticResource TextBox.Static.Border}"/>
-    <Setter Property="Foreground" Value="{DynamicResource {x:Static SystemColors.ControlTextBrushKey}}"/>
-    <Setter Property="BorderThickness" Value="1"/>
-    <Setter Property="KeyboardNavigation.TabNavigation" Value="None"/>
-    <Setter Property="HorizontalContentAlignment" Value="Left"/>
-    <Setter Property="FocusVisualStyle" Value="{x:Null}"/>
-    <Setter Property="AllowDrop" Value="true"/>
-    <Setter Property="HorizontalScrollBarVisibility" Value="Auto"/>
-    <Setter Property="ScrollViewer.PanningMode" Value="VerticalFirst"/>
-    <Setter Property="Stylus.IsFlicksEnabled" Value="False"/>
-    <Setter Property="Template">
-      <Setter.Value>
-        <ControlTemplate TargetType="{x:Type TextBoxBase}">
-          <Border x:Name="border" Background="{TemplateBinding Background}" SnapsToDevicePixels="True"
-                  BorderBrush="{TemplateBinding BorderBrush}" BorderThickness="{TemplateBinding BorderThickness}">
-            <DockPanel>
-              <Button DockPanel.Dock="Right" x:Name="btnClick" Content="Click Me" VerticalContentAlignment="Center"
-                      Height="{TemplateBinding Height}" Width="80"/>
-              <ScrollViewer x:Name="PART_ContentHost" Focusable="false" VerticalAlignment="Center"
-                            HorizontalScrollBarVisibility="Hidden" VerticalScrollBarVisibility="Hidden"/>
-            </DockPanel>
-          </Border>
-          <ControlTemplate.Triggers>
-            <Trigger Property="IsEnabled" Value="false">
-              <Setter Property="Opacity" TargetName="border" Value="0.56"/>
-            </Trigger>
-            <Trigger Property="IsMouseOver" Value="true">
-              <Setter Property="BorderBrush" TargetName="border" Value="{StaticResource TextBox.MouseOver.Border}"/>
-            </Trigger>
-            <Trigger Property="IsKeyboardFocused" Value="true">
-              <Setter Property="BorderBrush" TargetName="border" Value="{StaticResource TextBox.Focus.Border}"/>
-            </Trigger>
-          </ControlTemplate.Triggers>
-        </ControlTemplate>
-      </Setter.Value>
-    </Setter>
-
-    <Style.Triggers>
-      <MultiTrigger>
-        <MultiTrigger.Conditions>
-          <Condition Property="IsInactiveSelectionHighlightEnabled" Value="true"/>
-          <Condition Property="IsSelectionActive" Value="false"/>
-        </MultiTrigger.Conditions>
-        <Setter Property="SelectionBrush" Value="{DynamicResource {x:Static SystemColors.InactiveSelectionHighlightBrushKey}}"/>
-      </MultiTrigger>
-    </Style.Triggers>
-  </Style>
-
-</ResourceDictionary>
