+++ src/WPFCustomControlLib/CustomControl1.cs
@@ -1,49 +1,35 @@
-﻿using System.Text;
+﻿using System.Diagnostics.CodeAnalysis;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
 namespace WPFCustomControlLib;
 
-/// <summary>
-/// Follow steps 1a or 1b and then 2 to use this custom control in a XAML file.
-///
-/// Step 1a) Using this custom control in a XAML file that exists in the current project.
-/// Add this XmlNamespace attribute to the root element of the markup file where it is 
-/// to be used:
-///
-///     xmlns:MyNamespace="clr-namespace:WPFCustomControlLib"
-///
-///
-/// Step 1b) Using this custom control in a XAML file that exists in a different project.
-/// Add this XmlNamespace attribute to the root element of the markup file where it is 
-/// to be used:
-///
-///     xmlns:MyNamespace="clr-namespace:WPFCustomControlLib;assembly=WPFCustomControlLib"
-///
-/// You will also need to add a project reference from the project where the XAML file lives
-/// to this project and Rebuild to avoid compilation errors:
-///
-///     Right click on the target project in the Solution Explorer and
-///     "Add Reference"->"Projects"->[Select this project]
-///
-///
-/// Step 2)
-/// Go ahead and use your control in the XAML file.
-///
-///     <MyNamespace:CustomControl1/>
-///
-/// </summary>
-public class CustomControl1 : Control
+public class CustomControl1 : TextBox
 {
-    static CustomControl1()
-    {
-        DefaultStyleKeyProperty.OverrideMetadata(typeof(CustomControl1), new FrameworkPropertyMetadata(typeof(CustomControl1)));
-    }
+	static CustomControl1()
+	{
+		DefaultStyleKeyProperty.OverrideMetadata( typeof( CustomControl1 ),
+			new FrameworkPropertyMetadata( typeof( CustomControl1 ) ) );
+	}
+
+	private Button? _btnClick = null;
+
+	public override void OnApplyTemplate()
+	{
+		base.OnApplyTemplate();
+
+		_btnClick = GetTemplateChild( "btnClick" ) as Button;
+		if( _btnClick is not null )
+		{
+			_btnClick.Click += btnClick_OnClick;
+		}
+	}
+
+	private int counter;
+	[SuppressMessage( "Style", "IDE1006:Naming Styles", Justification = "Control name" )]
+	private void btnClick_OnClick( object sender, RoutedEventArgs e )
+	{
+		counter++;
+		Text = counter.ToString();
+	}
 }
