+++ src/WPFCustomControlLib/Themes/Button.xaml
@@ -1,0 +1,46 @@
+﻿<ResourceDictionary
+  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+  <!-- Theme colors -->
+  <Color x:Key="themeButtonBackgroundColor">#FFBEE6FD</Color>
+  <Color x:Key="themeButtonBorderColor">#FF569DE5</Color>
+  <Color x:Key="themeButtonMouseOverColor">LightSkyBlue</Color>
+
+  <SolidColorBrush x:Key="buttonBackgroundBrush" Color="{DynamicResource themeButtonBackgroundColor}"/>
+  <SolidColorBrush x:Key="buttonBorderBrush" Color="{DynamicResource themeButtonBorderColor}"/>
+  <SolidColorBrush x:Key="buttonMouseOverBrush" Color="{DynamicResource themeButtonMouseOverColor}"/>
+
+  <!-- Button brushes -->
+  <SolidColorBrush x:Key="Button.Disabled.Background" Color="#FFF4F4F4"/>
+  <SolidColorBrush x:Key="Button.Disabled.Border" Color="#FFADB2B5"/>
+  <SolidColorBrush x:Key="Button.Disabled.Foreground" Color="#FF838383"/>
+
+  <ControlTemplate x:Key="commonButtonTemplate" TargetType="{x:Type ButtonBase}">
+    <Border x:Name="bd" CornerRadius="3" BorderThickness="{TemplateBinding BorderThickness}"
+            Background="{DynamicResource buttonBackgroundBrush}"
+            BorderBrush="{DynamicResource buttonBorderBrush}">
+      <ContentPresenter x:Name="cp" TextElement.Foreground="{TemplateBinding Foreground}"
+                        HorizontalAlignment="{TemplateBinding HorizontalContentAlignment}"
+                        VerticalAlignment="{TemplateBinding VerticalContentAlignment}"/>
+    </Border>
+    <ControlTemplate.Triggers>
+      <Trigger Property="IsMouseOver" Value="True">
+        <Setter TargetName="bd" Property="Background" Value="{DynamicResource buttonMouseOverBrush}"/>
+      </Trigger>
+      <Trigger Property="IsEnabled" Value="false">
+        <Setter Property="Background" TargetName="bd" Value="{StaticResource Button.Disabled.Background}"/>
+        <Setter Property="BorderBrush" TargetName="bd" Value="{StaticResource Button.Disabled.Border}"/>
+        <Setter Property="TextElement.Foreground" TargetName="cp" Value="{StaticResource Button.Disabled.Foreground}"/>
+      </Trigger>
+    </ControlTemplate.Triggers>
+  </ControlTemplate>
+
+  <Style x:Key="commonButtonStyle" TargetType="{x:Type Button}">
+    <Setter Property="MinWidth" Value="80"/>
+    <Setter Property="Margin" Value="5"/>
+    <Setter Property="VerticalContentAlignment" Value="Center"/>
+    <Setter Property="Template" Value="{StaticResource commonButtonTemplate}"/>
+  </Style>
+
+</ResourceDictionary>
