+++ src/WPFCustomControlLib/Themes/ListView.xaml
@@ -1,0 +1,16 @@
+﻿<ResourceDictionary
+  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+  <Style x:Key="commonListViewStyle" TargetType="{x:Type ListView}">
+    <Setter Property="Background" Value="GhostWhite"/>
+    <Setter Property="SelectionMode" Value="Single"/>
+  </Style>
+
+  <Style x:Key="commonGridColumnHeaderStyle" TargetType="{x:Type GridViewColumnHeader}">
+    <Setter Property="HorizontalContentAlignment" Value="Left"/>
+    <Setter Property="FontWeight" Value="Bold"/>
+    <Setter Property="Margin" Value="5,0,0,0"/>
+  </Style>
+
+</ResourceDictionary>
